Changing authentication, starting transfer to SQL database
