+++ main.go
@@ -1,9 +1,6 @@
 package main
 
 import (
-	"fmt"
-	"log"
-
 	"todo-list/web-service-gin/docs"
 
 	"github.com/gin-gonic/gin"
@@ -22,28 +19,6 @@ import (
 func main() {
 	initSQL()
 
-	tskID, err := addTask(Task{
-		ID:       0,
-		Title:    "Go to the gym",
-		Complete: "false",
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Printf("ID of added task: %v\n", tskID)
-
-	tasks, err := tasksByComplete("false")
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Printf("Tasks found: %v\n", tasks)
-
-	tsk, err := taskByID(1)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Printf("Task found: %v\n", tsk)
-
 	router := gin.Default()
 
 	docs.SwaggerInfo.BasePath = ""
@@ -52,7 +27,7 @@ func main() {
 	router.GET("/health", getHealth)
 	router.GET("/tasks", getTasks)
 	router.GET("/tasks/:id", getTaskByID)
-	router.POST("/tasks", postTasks)
+	router.POST("/tasks", postTask)
 	router.PUT("/tasks/:id", putTasks)
 	router.DELETE("/tasks/:id", deleteTask)
